API

CellularPotts.jl includes the following core functions.

Index

CellularPotts.AdhesionPenalty

CellularPotts.CellPotts

CellularPotts.CellTable

CellularPotts.MigrationPenalty

CellularPotts.Penalty

CellularPotts.PerimeterPenalty

CellularPotts.VolumePenalty

CellularPotts.addCellProperty

CellularPotts.addNewCell

CellularPotts.countcells

CellularPotts.countcelltypes

CellularPotts.newCellState

CellularPotts.removeCell

Full docs

CellularPotts.AdhesionPenalty — Type

AdhesionPenalty(J::Matrix{Int})

An concrete type that penalizes neighboring grid locations from different cells.

Requires a symmetric matrix J where J[n,m] gives the adhesion penality for cells with types n and m. J is zero-indexed meaning J[0,1] and J[1,0] corresponds to the :Medium ↔ :Cell1 adhesion penalty.

Note: J is automatically transformed to be a zero-indexed offset array.

CellularPotts.CellPotts — Type

CellPotts(space, initialCellState, penalties)

A data container that holds information to run the cellular potts simulation.

Requires three inputs:

space: a region where cells can exist, generated using CellSpace().

initialCellState: a table where rows are cells and columns are cell properties, generated using newCellState().

penalties: a vector of penalties to append to the model.

CellularPotts.CellTable — Type

CellTable

An concrete type that stores a table where each row is a cell and each column is a cell property.

This table can be genereated using the newCellState() function.

CellularPotts.MigrationPenalty — Type

MigrationPenalty(maxAct, λ, gridSize)

An concrete type that encourages cells to protude and drag themselves forward.

Two integer parameters control how cells protude:

maxAct: A maximum activity a grid location can have

λ: A parameter that controls the strength of this penalty

Increasing maxAct will cause grid locations to more likely protrude. Increasing λ will cause those protusions to reach farther away.

MigrationPenalty also requires a list of cell types to apply the penalty to and the grid size (Space.gridSize).

CellularPotts.Penalty — Type

Penalty

An abstract type representing a constraint imposed onto the cellular potts model.

To add a new penalty, a new struct subtyping Penalty needs to be defined and the addPenalty!() function needs to be extended to include the new penalty.

Note: variables associated with a new penalty may need to be offset such that index 0 maps to :Medium, index 1 maps to :Cell1, etc.

CellularPotts.PerimeterPenalty — Type

PerimeterPenalty(λᵥ::Vector{Int})

An concrete type that penalizes cells that deviate from their desired perimeter.

Requires a vector λₚ with n penalties where n is the number of cell types. λₚ is zero-indexed meaning λₚ[0] corresponds to the :Medium perimeter penalty (which is set to zero).

Note: λₚ is automatically transformed to be a zero-indexed offset array and does not require the perimeter penalty for :Medium.

CellularPotts.VolumePenalty — Type

VolumePenalty(λᵥ::Vector{Int})

An concrete type that penalizes cells that deviate from their desired volume.

Requires a vector λᵥ with n penalties where n is the number of cell types. λᵥ is zero-indexed meaning λᵥ[0] corresponds to the :Medium volume penalty (which is set to zero).

Note: λᵥ is automatically transformed to be a zero-indexed offset array and does not require the volume penalty for :Medium.

CellularPotts.addCellProperty — Method

addCellProperty(df::CellTable, propertyName, defaultValue, cellName)
addCellProperty(df::CellTable, propertyName, values)

Given a cellTable, add a new column called propertyName with a given default value.

A cellTable is generated using the newCellState() function.

cellName can be the name of one cell type or a vector of cell types. Cells not included in cellName will be given a value of missing.

If cellNames are not specified, a vector of values can be supplied for every cell type.

CellularPotts.addNewCell — Method

addNewCell(df::CellTable, cell<:NamedTuple)

Given a cellTable, add a new row corresponding to a new cell in the mode. Property names in the for the cell need to match column names in the cellTable

CellularPotts.countcells — Method

countcells(cpm::CellPotts)
countcells(df::CellTable)

Count the number of cells in the model

CellularPotts.countcelltypes — Method

countcelltypes(cpm::CellPotts)
countcelltypes(df::CellTable)

Count the number of cell types in the model

CellularPotts.newCellState — Method

newCellState(names, volumes, counts)

Create a new cellTable where each row corresponds to a cell.

By default, this function generates a table with the following columns:

names::Vector{Symbol}: List of names given to cells (e.g. :TCell)

cellIDs::Vector{<:Integer}: A unqiue number given to a cell

typeIDs::Vector{<:Integer}: A number corresponding to the cell's name

volumes::Vector{<:Integer}: Number of grid squares occupied

desiredVolumes::Vector{<:Integer}: Desired number of grid square

perimeters::Vector{<:Integer}: Cell border penality

desiredPerimeters::Vector{<:Integer}: Desired cell border penality

The first row in the table is reserved for :Medium which is the name given to grid locations not belonging to any cell and is given an index of 0 (The first cell is given an index of 1).

Of note, desiredPerimeters are calculated as the minimal perimeter given the cell's volume.

CellularPotts.removeCell — Method

removeCell(df::CellTable, cellID)

Given a cellTable, remove the cell with provided cellID.